Move UnnestNode to SPI and add to collector optimizer
