+++ presto-main-base/src/main/java/com/facebook/presto/cost/UnnestStatsRule.java
@@ -16,10 +16,10 @@ package com.facebook.presto.cost;
 import com.facebook.presto.Session;
 import com.facebook.presto.cost.ComposableStatsCalculator.Rule;
 import com.facebook.presto.matching.Pattern;
+import com.facebook.presto.spi.plan.UnnestNode;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
 import com.facebook.presto.sql.planner.TypeProvider;
 import com.facebook.presto.sql.planner.iterative.Lookup;
-import com.facebook.presto.sql.planner.plan.UnnestNode;
 
 import java.util.List;
 import java.util.Map;
