+++ presto-main-base/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -67,11 +67,6 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitUnnest(UnnestNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitGroupId(GroupIdNode node, C context)
     {
         return visitPlan(node, context);
