+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -149,4 +149,9 @@ public abstract class PlanVisitor<R, C>
     {
         return visitPlan(node, context);
     }
+
+    public R visitUnnest(UnnestNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
 }
